--- test/test.js
@@ -1,22 +1,0 @@
-var tape   = require('tape')
-var leveljs = require('../')
-var testCommon = require('./util/test-common')
-
-// load IndexedDBShim in the tests
-// require('./util/idb-shim.js')()
-
-/*** compatibility with basic LevelDOWN API ***/
-require('abstract-leveldown/abstract/leveldown-test').args(leveljs, tape, testCommon)
-require('abstract-leveldown/abstract/open-test').open(leveljs, tape, testCommon)
-require('abstract-leveldown/abstract/put-test').all(leveljs, tape, testCommon)
-require('abstract-leveldown/abstract/del-test').all(leveljs, tape, testCommon)
-require('abstract-leveldown/abstract/get-test').all(leveljs, tape, testCommon)
-require('abstract-leveldown/abstract/put-get-del-test').all(leveljs, tape, testCommon)
-require('abstract-leveldown/abstract/batch-test').all(leveljs, tape, testCommon)
-require('abstract-leveldown/abstract/chained-batch-test').all(leveljs, tape, testCommon)
-require('abstract-leveldown/abstract/close-test').close(leveljs, tape, testCommon)
-require('abstract-leveldown/abstract/iterator-test').all(leveljs, tape, testCommon)
-require('abstract-leveldown/abstract/iterator-range-test').all(leveljs, tape, testCommon)
-
-// non abstract-leveldown tests:
-require('./custom-tests.js').all(leveljs, tape, testCommon)
